+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Added
+
+* Support custom Activity vocabulary extensions via `additionalContext` config option
+
 ## v3.4.0 (2022-09-16)
 
 **Urgent update**: Your SSL certificates may not be able to renew