+++ CHANGELOG.md
@@ -8,6 +8,7 @@ identities throughout the metaverse.
 [More info](./doc/ControlledAccounts.md#logging-in-a-controlled-account)
   * Configuration option to redirect to a custom login page for controlled accounts [More info](./doc/ControlledAccounts.md#custom-login-redirect)
 * Configuration option to alter session cookie id
+* Support custom Activity vocabulary extensions via `additionalContext` config option
 
 ## v3.4.0 (2022-09-16)